+++ R/mark.R
@@ -4,7 +4,13 @@
 #' 
 #' @inheritParams e_bar
 #' @param serie Serie or vector of series to mark on, defaults to all series.
-#' @param data Placement.
+#' @param data Placement of point, line or area.
+#' @param title A convenience argument to easily set label, see details.
+#' 
+#' @details To set a label you need to either use the \code{title} 
+#' argument or pass a list specifying the label formatter.
+#' \code{label = list(formatter = "label")}. The former is more convenient
+#' but more limited, e.g.: you cannot specify the placement of the label.
 #' 
 #' @examples 
 #' max <- list(
@@ -40,7 +46,7 @@
 #' 
 #' @rdname mark
 #' @export
-e_mark_point <- function(e, serie = NULL, data = NULL, ...){
+e_mark_point <- function(e, serie = NULL, data = NULL, ..., title = NULL){
   
   if(missing(e))
     stop("must pass e", call. = FALSE)
@@ -55,11 +61,15 @@ e_mark_point <- function(e, serie = NULL, data = NULL, ...){
     
     if(!is.null(data))
       point$data <- list(data)
+
+    if(!is.null(title) && !is.null(data)){
+      point$data[[1]]$label <- list(formatter = title)
+    }
     
     if(is.null(e$x$opts$series[[i]]$markPoint))
       e$x$opts$series[[i]]$markPoint <- append(e$x$opts$series[[i]]$markPoint, point)
     else
-      e$x$opts$series[[i]]$markPoint$data <- append(e$x$opts$series[[i]]$markPoint$data, list(data))
+      e$x$opts$series[[i]]$markPoint$data <- append(e$x$opts$series[[i]]$markPoint$data, point$data)
   }
   
   e
@@ -67,7 +77,7 @@ e_mark_point <- function(e, serie = NULL, data = NULL, ...){
 
 #' @rdname mark
 #' @export
-e_mark_line <- function(e, serie = NULL, data = NULL, ...){
+e_mark_line <- function(e, serie = NULL, data = NULL, ..., title = NULL){
   
   if(missing(e))
     stop("must pass e", call. = FALSE)
@@ -82,11 +92,14 @@ e_mark_line <- function(e, serie = NULL, data = NULL, ...){
     
     if(!is.null(data))
       point$data <- list(data)
+
+    if(!is.null(title) && !is.null(data))
+      point$data[[1]]$label <- list(formatter = title)
     
     if(is.null(e$x$opts$series[[i]]$markLine))
       e$x$opts$series[[i]]$markLine <- append(e$x$opts$series[[i]]$markLine, point)
     else
-      e$x$opts$series[[i]]$markLine$data <- append(e$x$opts$series[[i]]$markLine$data, list(data))
+      e$x$opts$series[[i]]$markLine$data <- append(e$x$opts$series[[i]]$markLine$data, point$data)
   }
   
   e
@@ -94,7 +107,7 @@ e_mark_line <- function(e, serie = NULL, data = NULL, ...){
 
 #' @rdname mark
 #' @export
-e_mark_area <- function(e, serie = NULL, data = NULL, ...){
+e_mark_area <- function(e, serie = NULL, data = NULL, ..., title = NULL){
   
   if(missing(e))
     stop("must pass e", call. = FALSE)
@@ -109,11 +122,14 @@ e_mark_area <- function(e, serie = NULL, data = NULL, ...){
     
     if(!is.null(data))
       point$data <- list(data)
+
+    if(!is.null(title) && !is.null(data))
+      point$data[[1]]$label <- list(formatter = title)
     
     if(is.null(e$x$opts$series[[i]]$markArea))
       e$x$opts$series[[i]]$markArea <- append(e$x$opts$series[[i]]$markArea, point)
     else
-      e$x$opts$series[[i]]$markArea$data <- append(e$x$opts$series[[i]]$markArea$data, list(data))
+      e$x$opts$series[[i]]$markArea$data <- append(e$x$opts$series[[i]]$markArea$data, point$data)
   }
   
   e