+++ docs/articles/mark_files/echarts4r-binding-0.3.0.9000/echarts4r.js
@@ -45,7 +45,6 @@ HTMLWidgets.widget({
         if(x.draw === true)
           chart.setOption(opts);
         
-        
         // shiny callbacks
         if (HTMLWidgets.shinyMode) {